[File structure changes/ Human added to Object Designator/ some Code changes in receptionist tests and Fluent implemented]
