+++ demos/pycram_receptionist_demo/utils/misc.py
@@ -7,26 +7,20 @@ from std_msgs.msg import String
 from demos.pycram_receptionist_demo.deprecated import talk_actions
 from deprecated import deprecated
 
+pub_nlp = rospy.Publisher('/startListener', String, queue_size=10)
 
 
-@deprecated(reason="Newst version uses the knowledge interface")
-def talk_request(data: String):
+def talk_request(name_drink: List):
     """
     callback function that takes the data from nlp (name and drink) and lets the robot talk
-    :param data: String "name drink"
+    :param name_drink: List ["name", "drink"]
     """
-
-    name_drink = data.data.split(" ")
-    talk_actions.name_drink_talker(name_drink)
     rospy.loginfo("nlp data:" + name_drink[0] + " " + name_drink[1])
 
-    rospy.loginfo("stop looking now")
-    giskardpy.stop_looking()
-    rospy.loginfo("Navigating now")
-    NavigateAction([Pose([3, 5, 0], [0, 0, 1, 1])]).resolve().perform()
 
 
-@deprecated(reason="Newst version uses the knowledge interface")
+
+
 def talk_error(data):
     """
     callback function if no name/drink was heard
@@ -44,5 +38,5 @@ def introduce(name1, drink1, name2, drink2):
     """
     first = "Hey" + name2 + " This is " + name1 + "and the favorite drink of your guest is " + drink1
     second = name1 + "This is " + name2 + "his favorite drink is " + drink2
-    TalkingMotion(first)
-    TalkingMotion(second)
+    TalkingMotion(first).resolve.perform()
+    TalkingMotion(second).resolve.perform()
